third_party/WebKit/LayoutTests/platform/win/http/tests/security/suborigins/suborigin-cookies-expected.txt

This is a testharness.js-based test.
PASS Document is cookie-averse 
PASS Document getters and setters still work 
PASS Cookies set in a frame with a regular, same-origin src do not modify the suborigin's document.cookie 
PASS Cookies set in an about:blank frame do not modify the suborigin's document.cookie and also have an empty document.cookie 
PASS Cookies set in a srcdoc frame do not modify the suborigin's document.cookie and also have an empty document.cookie 
Harness: the test ran to completion.
